requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.7.1
pydantic-core==2.18.2
typing-extensions>=4.9.0
annotated-types>=0.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==3.2.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
jinja2==3.1.3
itsdangerous==2.1.2
email-validator==2.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
